dev-requirements.txt
pytest-django
pytest-randomly
coverage
